requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-filter==25.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
python-dotenv==1.1.0
dj-database-url==2.3.0
